erl_call: Lift in erl_start.h and erl_start.c
Lift in erl_start.h and erl_start.c into erl_call.c and remove
erl_start.h and erl_start.c to make erl_call.c self contained.
